pyproject.toml
[tool.poetry]
name = "tide"
version = "0.1.0"
description = ""
authors = ["martins0n <33594071+martins0n@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
torch = "<2.0.0"
pandas = ">=1.1,<2"
seaborn = ">=0.11.0,<1"
wandb = "^0.15.0"
etna = {version = "2.0.0", extras = ["torch"]}
pyarrow = "^11.0.0"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
